feat: Refactor AuthController registration logic; streamline UserModel creation and validation feat: Update SiteController contact method to return UserModel instances directly fix: Ensure ErrorHandler sets HTTP response status as an integer refactor: Enhance Router dispatch method with match expression for cleaner callback handling refactor: Modify ActiveRecord fill method to accept mixed data types feat: Update UserModel rules to enforce unique username and email validation

+++ app/core/ErrorHandler.php
@@ -15,7 +15,7 @@ class ErrorHandler
 
     public static function handleException(Throwable $e)
     {
-        $status = $e->getCode() ?: 500;
+        $status = (int) ($e->getCode() ?: 500);
         http_response_code($status);
 
         $exceptionClass = get_class($e);
